requirements.txt
GitPython >= 3.0.0